+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.1.0
+
+*2019-02-01*
+
+- Add ability to specify fixture files only in `preprocessors` configuration,
+  removing the need to also specify them in the `files` configuration.
+- Added plugin tests.
+
 ## 1.0.1
 
 *2019-01-08*